Recover skill costs and battle fade records

+++ src/promoted/code1_0011.c
@@ -54,7 +54,7 @@ extern void func_00453570(void);
 extern void func_0043f9c8(u8 *arg0, s32 arg1, s32 arg2);
 extern void func_00452080(void);
 extern s32 func_0010f560(s16 arg0, u16 arg1);
-extern s32 func_0010f600(s16 arg0, u16 arg1);
+extern u32 func_0010f600(s16 arg0, s32 arg1);
 extern void *func_00243840(u16 arg0);
 
 extern void func_0046d3b0(s32 parent, s32 arg0, s32 arg1, f32 x, f32 y,
